requirements.txt
Flask>=3.0.0
APScheduler>=3.10.4
gspread>=6.0.0
google-auth>=2.28.0
twilio>=9.0.0
pytz>=2023.3
gunicorn>=21.2.0
